Fix bug, add docstrings and readme, tidy files.
