+++ .idea/scikit_maad.iml
@@ -2,9 +2,13 @@
 <module type="PYTHON_MODULE" version="4">
   <component name="NewModuleRootManager">
     <content url="file://$MODULE_DIR$">
+      <excludeFolder url="file://$MODULE_DIR$/.venv" />
+      <excludeFolder url="file://$MODULE_DIR$/notes" />
+      <excludeFolder url="file://$MODULE_DIR$/temp_output" />
+      <excludeFolder url="file://$MODULE_DIR$/test_data" />
       <excludeFolder url="file://$MODULE_DIR$/venv" />
     </content>
-    <orderEntry type="jdk" jdkName="Python 3.7 (scikit_maad_biosound)" jdkType="Python SDK" />
+    <orderEntry type="jdk" jdkName="Python 3.12 (scikit-maad-biosound)" jdkType="Python SDK" />
     <orderEntry type="sourceFolder" forTests="false" />
   </component>
 </module>
